+++ src/Proxy/SharedProxy.php
@@ -20,6 +20,15 @@ use Microsoft\Graph\Model\Shared;
 /**
  * A proxy to a \Microsoft\Graph\Model\Shared instance.
  *
+ * @property-read \Krizalys\Onedrive\Proxy\IdentitySetProxy $owner
+ *                The owner.
+ * @property-read string $scope
+ *                The scope.
+ * @property-read \Krizalys\Onedrive\Proxy\IdentitySetProxy $sharedBy
+ *                The shared by.
+ * @property-read \DateTime $sharedDateTime
+ *                The shared date/time.
+ *
  * @since 2.0.0
  *
  * @api
@@ -41,5 +50,40 @@ class SharedProxy extends EntityProxy
     public function __construct(Graph $graph, Shared $shared)
     {
         parent::__construct($graph, $shared);
+    }
+
+    /**
+     * Getter.
+     *
+     * @param string $name
+     *        The name.
+     *
+     * @return mixed
+     *         The value.
+     *
+     * @since 2.5.0
+     */
+    public function __get($name)
+    {
+        $shared = $this->entity;
+
+        switch ($name) {
+            case 'owner':
+                $owner = $shared->getOwner();
+                return $owner !== null ? new IdentitySetProxy($this->graph, $owner) : null;
+
+            case 'scope':
+                return $shared->getScope();
+
+            case 'sharedBy':
+                $sharedBy = $shared->getSharedBy();
+                return $sharedBy !== null ? new IdentitySetProxy($this->graph, $sharedBy) : null;
+
+            case 'sharedDateTime':
+                return $shared->getSharedDateTime();
+
+            default:
+                return parent::__get($name);
+        }
     }
 }